Code improvements

+++ src/Controllers/PermissionController.php
@@ -3,6 +3,7 @@
 namespace marsoltys\uservel\Controllers;
 
 use Illuminate\Http\Request;
+use SCC\EndeavourCMS\Support\Permissions;
 use Spatie\Permission\Models\Permission;
 
 class PermissionController extends Controller
@@ -14,6 +15,7 @@ class PermissionController extends Controller
      */
     public function index()
     {
+        $this->authorize(Permissions::PERMISSION_VIEW);
         $permissions = Permission::orderBy('name')->get();
         return view('uservel::permission.list',[
             'permissions' => $permissions
@@ -27,6 +29,7 @@ class PermissionController extends Controller
      */
     public function create()
     {
+        $this->authorize(Permissions::PERMISSION_CREATE);
         return view('uservel::permission.form', [
             'title' => 'Create permission'
         ]);
@@ -40,6 +43,7 @@ class PermissionController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize(Permissions::PERMISSION_CREATE);
         $data = $request->validate([
             'name' => 'required|unique:permissions|max:255'
         ]);
@@ -67,6 +71,7 @@ class PermissionController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize(Permissions::PERMISSION_EDIT);
         $permission = Permission::findOrFail($id);
         return view('uservel::permission.form', [
             'permission'=>$permission,
@@ -83,6 +88,7 @@ class PermissionController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $this->authorize(Permissions::PERMISSION_EDIT);
         $permission = Permission::findOrFail($id);
 
         $data = $request->validate([
@@ -102,6 +108,7 @@ class PermissionController extends Controller
      */
     public function destroy(Request $request, $id)
     {
+        $this->authorize(Permissions::PERMISSION_DELE);
         if (Permission::destroy($id)) {
             $request->session()->flash('laralert', [[
                 'type' => 'success',
